Fix compiler error

+++ src/battle_dynamax.c
@@ -485,7 +485,7 @@ static u32 GetMaxMoveStatusEffect(u32 move)
         }
         case MAX_EFFECT_EFFECT_SPORE_FOES:
         {
-            static const u8 sBefuddleEffects[] = {STATUS1_PARALYSIS, STATUS1_POISON, B_USE_DROWSY == TRUE ? STATUS1_DROWSY : STATUS1_SLEEP};
+            static const u16 sBefuddleEffects[] = {STATUS1_PARALYSIS, STATUS1_POISON, (B_USE_DROWSY == TRUE ? STATUS1_DROWSY : STATUS1_SLEEP)};
             return RandomElement(RNG_G_MAX_BEFUDDLE, sBefuddleEffects);
         }
         // Status 2
